# ENH: improve scipy.interpolate.RegularGridInterpolator performance