#### What does this implement/fix?

This PR reimplements the `scipy.interpolate.RegularGridInterpolator._evaluate_linear` method to be slightly more performant.

To compute linear interpolation over an n-dimensional grid, the `_evaluate_linear` method iterates over the vertices of an n-d hypercube. The previous implementation uses `np.where` to determine the "weights" to attach to the grid values at a vertex. These weights can instead be directly inferred from the vertex iteration order. The proposed implementation included here takes this approach, eliminating the unnecessary call to `np.where`. This change gives a nontrivial speedup in most situations.

#### Additional information

I wrote a set of benchmarks specifically for the `RegularGridInterpolator`. Below are the results of running the benchmarks locally.

*Previous implementation*

```
              --                                          n_samples
              ----------------------- --------------------------------------------------
               ndim   max_coord_size       10        100          1000         10000
              ====== ================ =========== ========== ============= =============
                2           10          177±7μs    213±20μs     311±40μs    1.03±0.03ms
                2           40         160±0.8μs   197±9μs      338±10μs    1.31±0.05ms
                2          200          177±7μs    184±9μs      356±20μs     1.89±0.2ms
                3           10          291±20μs   318±10μs     462±30μs     1.88±0.1ms
                3           40          309±10μs   339±20μs     589±20μs     2.29±0.2ms
                3          200          310±7μs    311±3μs      678±70μs     3.26±0.1ms
                4           10          544±30μs   566±10μs    1.05±0.1ms    3.80±0.3ms
                4           40          601±40μs   601±20μs    976±100μs     4.10±0.7ms
                4          200          661±10μs   744±50μs   1.38±0.03ms    9.26±0.3ms
```

*New implementation this PR aims to merge*

```
              --                                        n_samples
              ----------------------- ----------------------------------------------
               ndim   max_coord_size      10        100        1000        10000
              ====== ================ ========== ========== ========== =============
                2           10         142±8μs    151±4μs    229±5μs      907±20μs
                2           40         136±10μs   153±30μs   260±8μs     1.20±0.1ms
                2          200         146±9μs    152±7μs    289±20μs    1.54±0.2ms
                3           10         239±40μs   255±30μs   401±80μs    1.58±0.3ms
                3           40         204±10μs   223±10μs   401±80μs    1.85±0.2ms
                3          200         205±20μs   262±30μs   533±60μs    2.63±0.1ms
                4           10         311±20μs   316±9μs    565±20μs   2.36±0.06ms
                4           40         298±20μs   354±20μs   687±20μs    3.75±0.3ms
                4          200         395±30μs   423±10μs   910±70μs    7.65±0.4ms
```