@Kai-Striega 2022-06-23
> I haven't noticed that particular issue for my use cases. With the RegularGridInterpolator, I find that if the grid itself can fit into memory, then I can readily interpolator over it. 

I tried your example, it works fine for me too. The example in gh-7663 does not work for me, despite working with a smaller array. Weird. Maybe it has something to do with meshgrid? 

I'm trying to figure out why this benchmark runs so much quicker than my benchmark in gh-7663. Sorry to be a pain, but could you give that benchmark a go?

> If you do Cythonize the _evaluate_linear method, know that it does n * 2^n iterations in the double for-loop, where n is dimension of the underlying grid. I've found the slowest step is the fancy indexing (self.values[edge_indices] in the source code).

Thanks. I'll keep that in mind. I don't think there's much we can do to improve the n * 2^n part is there?