@zebengberg 2022-06-23
> I tried your example, it works fine for me too. The example in [gh-7663](https://github.com/scipy/scipy/issues/7663) does not work for me, despite working with a smaller array. Weird. Maybe it has something to do with meshgrid?
> 
> I'm trying to figure out why this benchmark runs so much quicker than my benchmark in [gh-7663](https://github.com/scipy/scipy/issues/7663). Sorry to be a pain, but could you give that benchmark a go?

I checked out your benchmarks in #7663 . I'm definitely seeing a memory explosion when `n` gets large. The RGI calls down to [`_ndim_coords_from_arrays`](https://github.com/scipy/scipy/blob/b4857872e7ccbcfa86ad06e001c10334843dc0b7/scipy/interpolate/interpnd.pyx#L157), which broadcasts the arrays `xg` and `yg` against each other. In other words, you'd see the same memory footprint as if you created `xg` and `yg` with `np.meshgrid(..., sparse=False)`.

I think the RGI is most useful when the input `xi` into the`__call__` method is somewhat small relative to the size of the underlying grid. This is the case with my example in the thread above ... there, the underlying grid has shape `(30000, 30000)`, whereas the `xi` only has one million entries ... small relative to the `30000 x 30000` size of the grid.

I also looked at the implementation of `interp2d`. I'm seeing a memory explosion when the [`regrid_smth`](https://github.com/scipy/scipy/blob/b4857872e7ccbcfa86ad06e001c10334843dc0b7/scipy/interpolate/src/fitpack.pyf#L630) subroutine is called. I don't really understand what's happening there, but for me, it chewed up 3x the memory of `z` grid itself. This may be causing your unexpected crashes.

> I don't think there's much we can do to improve the n * 2^n part is there?
Right ... that seems intrinsic to the algorithm!